Move helper display functions for displaying clubs to separate folder.
