+++ frontend/src/components/pages/your_clubs/ClubList.tsx
@@ -1,8 +1,11 @@
 import * as React from "react";
+import { BookInfo } from "../club_display/BookInfo";
 import { BookInterface } from "../../../services/backend_service_interface/backend_service_interface";
 import Box from "@material-ui/core/Box";
 import Button from "@material-ui/core/Button";
+import { ClubDescription } from "../club_display/ClubDescription";
 import { ClubInterface } from "../../../services/backend_service_interface/backend_service_interface";
+import { ContentWarnings } from "../club_display/ContentWarnings";
 import { createStyles } from "@material-ui/core/styles";
 import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
@@ -17,10 +20,6 @@ import { Theme } from "@material-ui/core/styles";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
-    boldTextElement: {
-      fontWeight: 'bold',
-      marginBottom: theme.spacing(0),
-    },
     button : {
       marginLeft: theme.spacing(1),
       marginTop: theme.spacing(1),
@@ -63,21 +62,6 @@ const useStyles = makeStyles((theme: Theme) =>
       display: 'flex',
       flexDirection: 'column',
       marginBottom: theme.spacing(3),
-    },
-    textElement: {
-      marginBottom: theme.spacing(0),
-      marginTop: theme.spacing(0),
-    },
-    warningsHeader: {
-      fontWeight: 'bold',
-      marginBottom: theme.spacing(0),
-      marginLeft: '-17px',
-    },
-    warningsList: {
-      fontWeight: 'bold',
-      justifyContent: 'left',
-      marginBottom: theme.spacing(0),
-      marginTop: theme.spacing(0),
     },
   }),
 );
@@ -179,72 +163,6 @@ export function ClubList(props: ClubListProps) {
     );
   } else {
     return (<div></div>);
-  }
-}
-
-interface ClubDescriptionProps {
-  description: string,
-};
-
-/** Displays the club's description. */
-function ClubDescription(props: ClubDescriptionProps) {
-  const classes = useStyles();
-
-  return (
-    <>
-      <div className={classes.boldTextElement}>
-        Description:
-      </div>
-      <p className={classes.textElement}>
-        {props.description}
-      </p>
-    </>
-  );
-}
-
-interface BookInfoProps {
-  book: BookInterface;
-};
-
-/** Displays the title and author of the club's current book. */
-function BookInfo(props: BookInfoProps) {
-  const classes = useStyles();
-
-  return (
-    <div>
-      <div className={classes.boldTextElement}>
-        Current Book:
-      </div>
-      <p className={classes.textElement}>
-        <div>{props.book.title}</div>
-        <div>by {props.book.author}</div>
-      </p>
-    </div>
-  );
-}
-
-/** Displays content warnings associated with the club's current book. */
-function ContentWarnings(props) {
-  const classes = useStyles();
-  const contentWarnings = props.contentWarnings;
-
-  if (contentWarnings) {
-    return (
-        <div>
-          <ul className={classes.warningsList}>
-            <div className={classes.warningsHeader}>Content Warnings:</div>
-            {contentWarnings.map((item, index) => (
-              <li>{item}</li>
-            ))}
-          </ul>
-        </div>
-    );
-  } else {
-    return (
-      <p className={classes.warningsList}>
-        No Content Warnings
-      </p>
-    );
   }
 }
